server/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
identify==2.6.1
nodeenv==1.9.1
platformdirs==4.3.6
pre-commit==3.8.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
ruff==0.6.8
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.6
